Refactor CreateService to FolderService
