+++ src/main/java/bio/terra/folder/service/folder/FolderService.java
@@ -1,13 +1,13 @@
-package bio.terra.folder.service.create;
+package bio.terra.folder.service.folder;
 
 import bio.terra.folder.db.FolderDao;
 import bio.terra.folder.generated.model.CreateFolderBody;
 import bio.terra.folder.generated.model.CreatedFolder;
-import bio.terra.folder.service.create.exception.InvalidNameException;
-import bio.terra.folder.service.create.exception.InvalidSpendProfileException;
-import bio.terra.folder.service.create.exception.NameConflictException;
-import bio.terra.folder.service.create.flight.FolderCreateFlight;
-import bio.terra.folder.service.create.flight.FolderFlightMapKeys;
+import bio.terra.folder.service.folder.exception.InvalidNameException;
+import bio.terra.folder.service.folder.exception.InvalidSpendProfileException;
+import bio.terra.folder.service.folder.exception.NameConflictException;
+import bio.terra.folder.service.folder.flight.FolderCreateFlight;
+import bio.terra.folder.service.folder.flight.FolderFlightMapKeys;
 import bio.terra.folder.service.iam.AuthenticatedUserRequest;
 import bio.terra.folder.service.job.JobBuilder;
 import bio.terra.folder.service.job.JobService;
@@ -15,15 +15,17 @@ import java.util.UUID;
 import java.util.regex.Pattern;
 import org.springframework.stereotype.Component;
 
+// Service for managing actions related to Folder CRUD operations. Operations for managing contained
+// objects belong in a separate service.
 @Component
-public class CreateService {
+public class FolderService {
 
   public static Pattern validNamePattern = Pattern.compile("^[\\w\\-\\s_]+$");
 
   private JobService jobService;
   private FolderDao folderDao;
 
-  public CreateService(JobService jobService, FolderDao folderDao) {
+  public FolderService(JobService jobService, FolderDao folderDao) {
     this.jobService = jobService;
     this.folderDao = folderDao;
   }
