+++ src/main/java/bio/terra/folder/service/folder/flight/FolderFlightMapKeys.java
@@ -1,4 +1,4 @@
-package bio.terra.folder.service.create.flight;
+package bio.terra.folder.service.folder.flight;
 
 public class FolderFlightMapKeys {
   private FolderFlightMapKeys() {}
